# internal/types/powergamer.go
package types

type PowerGamer struct {
	Rank     int    `json:"rank"`
	Name     string `json:"name"`
	Vocation string `json:"vocation"`
	Level    int    `json:"level"`
	Today    int    `json:"today"`
}